Quit being trigger happy when commiting.
CVS patchset: 5391
CVS date: 2003/09/16 02:08:03

+++ src/xine-utils/memcpy.c
@@ -381,7 +381,7 @@ static struct {
   char *name;
   void *(* function)(void *to, const void *from, size_t len);
 
-  unsigned long long time; /* This type could be used for non-MSC build too! */
+  uint64_t time; /* This type could be used for non-MSC build too! */
 
   uint32_t cpu_require;
 } memcpy_method[] =
